Python 3.10 support; convert to poetry/pyproject.toml
- add python 3.10 support

- update PGPy to the current version (necessary for python 3.10)

- deprecate python 3.4 and 3.5 support

- convert from setuptools to poetry

- fix parsing of DSC files with bad/missing inline signatures:
  we were depending on behavior of PGPy <=0.4.1 which was in fact
  a bug that was fixed in 0.4.2 and later.

- fix various pylint/flakehell/pycodestyle warnings

- apply `black` formatting everywhere

- split classes into modules

- convert from travis-ci to github actions since travis have
  decided to commit corporate suicide
