+++ .github/workflows/pr.yaml
@@ -1,0 +1,40 @@
+name: pull request
+
+on: [pull_request]
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version:
+        - "3.6"
+        - "3.7"
+        - "3.8"
+        - "3.9"
+        - "3.10"
+
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          python -m pip install --upgrade poetry
+          poetry install
+        env:
+          POETRY_VIRTUALENVS_CREATE: false
+      - name: Check formatting
+        run: poetry run black --check .
+      - name: Lint with flakehell
+        run: poetry run flakehell lint pydpkg test
+      - name: Lint with pylint
+        run: poetry run pylint -d R0912 -d W0511 pydpkg/
+      - name: Lint with pycodestyle (aka pep8)
+        run: poetry run pycodestyle --max-line-length=120 --ignore=E203 pydpkg/
+      - name: Test with pytest
+        run: poetry run pytest tests/
