--- .travis.yml
@@ -1,17 +1,0 @@
-dist: xenial
-language: python
-python:
-  - "3.4"
-  - "3.5"
-  - "3.6"
-  - "3.7"
-  - "3.8"
-  - "3.9"
-before_install:
-  - "pip install -U pip"
-install:
-  - "pip install -e .[test]"
-script:
-  - "py.test tests/"
-  - "pylint -d R0912 -d W0511 pydpkg/"
-  - "pep8 --max-line-length=90 --ignore=E203 pydpkg/"
