+++ tests/testdeb_1.1.1-bad.dsc.asc
@@ -24,11 +24,8 @@ Files:
  fc80e6e7f1c1a08b78a674aaee6c1548 232 testdeb_1.1.1-1.debian.tar.xz
 -----BEGIN PGP SIGNATURE-----
 
-iQEcBAEBCAAGBQJZSyL/AAoJEDBtfbmvfY+UH6EIALzdWgPraB/8xrzSum/g600+
-lPGrObE1vgtH0IdkWyBhLwZKZGSgeEI7U11EOC+z78ULTTDafjeXZDWTPhO+F23d
-VFSOTHnzhqTtoqa1lT4qL2A6kWCc3QxpFtI6dLuZtXFlEEKk9w2tu/GzSWtuzVKQ
-qNRDdLsiFz7kPwMWjvlaqVjz5wqtY3j5TpYpVH8uqtrCuzaycGjzTgMLxHa+o2T+
-AHYnZLRN22hLb1DeFaZDkNiC6qgx45FV98jh0sYTZA7A15MbwCvNLD0uws4ICwPl
-8ex3e0rf4FlpUuMNJTQmShuxaMGlbwtzKFy75SxQxxJ0r+7TL4/B37vBE9+BJqs=
-=Of1F
+FSOTHnzhqTtoqa1lT4qL2A6kWCc3QxpFtI6dLuZtXFlEEKk9w2tu/GzSWtuzVKQ
+NRDdLsiFz7kPwMWjvlaqVjz5wqtY3j5TpYpVH8uqtrCuzaycGjzTgMLxHa+o2T+
+ex3e0rf4FlpUuMNJTQmShuxaMGlbwtzKFy75SxQxxJ0r+7TL4/B37vBE9+BJqs=
+Of1F
 -----END PGP SIGNATURE-----
